Miscellaneous cleanup for the Android codebase: - update gradle plugins versions - cleanup java_godot_lib_jni
Note: logic was mostly moved around and no new logic/functionality was added.
